Updated the setup script
- Included Minio installation check, prompt and installation in talawa-api setup
- Updated the installation checks for MinIo
- Modularised the code to isolate the environment
